Code tweaks
